GAL-506 cleanups and fixups:  - wsrep API: changed wsrep_uuid_t aligner from uint64_t to size_t -    we need only host-level alignment here.  - fixed WSREP(this) issue properly by using WSREP_NNULL() instead.

+++ wsrep/wsrep_api.h
@@ -190,8 +190,8 @@ typedef enum wsrep_cb_status
  * UUID type - for all unique IDs
  */
 typedef union wsrep_uuid {
-    uint8_t  data[16];
-    uint64_t alignment;
+    uint8_t data[16];
+    size_t  alignment;
 } wsrep_uuid_t;
 
 /*! Undefined UUID */
